refactor(data_operations): remove dead code and consolidate duplicated logic (#259)
Dead code (verified, runtime-unreachable):
- Remove NullPolicy enum and reword the ~14 NullPolicy.* docstring/comment
  anchors that referenced it to neutral prose (Null-skip, Null-as-group,
  Nulls-last, Edge-null).
- Remove unused DataOperationsTestDataCreator.NULL_POSITIONS class var,
  the unused composite VARIED_EXPECTED dict, and the unused raw_data/
  sample_data pytest fixtures (plus orphaned imports) in the offset and
  window_aggregation test conftests.

Consolidate duplicated logic (behavior-preserving) into shared modules:
- Backend agg-function tables: pyarrow_agg_constants, polars_agg_constants,
  sqlite_agg_constants, duckdb_agg_constants. Honor the per-context
  divergences: DuckDB group FIRST/LAST vs window FIRST_VALUE/LAST_VALUE,
  Polars aggregation extends the shared 14 with first/last.
- PartitionByMixin._extract_partition_by shared by the ffill/ema/
  sessionization/resample bases.
- assert_pandas/polars/duckdb_source_col_present source-column guards and
  duckdb_drop_rn_restore row-order-restore helper.
- Move _assert_float_list_with_nulls onto DataOpsTestBase.

tox passes (pytest, ruff format/check, mypy --strict, bandit).
