+++ mloda/community/feature_groups/data_operations/row_preserving/ema/base.py
@@ -50,8 +50,10 @@ from mloda.core.abstract_plugins.components.feature_set import FeatureSet
 from mloda.core.abstract_plugins.components.options import Options
 from mloda.provider import DefaultOptionKeys, FeatureGroup
 
+from mloda.community.feature_groups.data_operations.base import PartitionByMixin
 
-class EmaFeatureGroup(FeatureChainParserMixin, FeatureGroup):
+
+class EmaFeatureGroup(PartitionByMixin, FeatureChainParserMixin, FeatureGroup):
     """Base class for exponential-moving-average operations that preserve row count."""
 
     PREFIX_PATTERN = r".*__ema_\d+$"
@@ -135,14 +137,6 @@ class EmaFeatureGroup(FeatureChainParserMixin, FeatureGroup):
         if span <= 0:
             raise ValueError(f"ema span must be a positive integer (span > 0), got {span} in {name!r}.")
         return span
-
-    @classmethod
-    def _extract_partition_by(cls, feature: Feature) -> list[str]:
-        """Return ``partition_by`` as a list (defaulting to ``[]`` when absent)."""
-        partition_by = feature.options.get(cls.PARTITION_BY)
-        if partition_by is None:
-            return []
-        return list(partition_by)
 
     @classmethod
     def _extract_order_by(cls, feature: Feature) -> str:
