+++ mloda/testing/feature_groups/data_operations/aggregation/aggregation.py
@@ -218,7 +218,7 @@ class AggregationTestBase(MaskTestMixin, DataOpsTestBase):
             assert result_map[region] == expected, f"region={region}"
 
     def test_null_policy_skip_avg_with_null_values(self) -> None:
-        """NullPolicy.SKIP: Group B has a null value_int at row 4. Avg should skip it."""
+        """Null-skip: Group B has a null value_int at row 4. Avg should skip it."""
         fs = make_feature_set("value_int__avg_agg", ["region"])
         result = self.implementation_class().calculate_feature(self.test_data, fs)
 
@@ -229,7 +229,7 @@ class AggregationTestBase(MaskTestMixin, DataOpsTestBase):
         assert result_map["B"] == pytest.approx(140.0 / 3.0, rel=1e-6)
 
     def test_null_policy_null_is_group(self) -> None:
-        """NullPolicy.NULL_IS_GROUP: Row 11 has region=None. It should form its own group."""
+        """Null-as-group: Row 11 has region=None. It should form its own group."""
         fs = make_feature_set("value_int__sum_agg", ["region"])
         result = self.implementation_class().calculate_feature(self.test_data, fs)
 
@@ -535,7 +535,7 @@ class AggregationTestBase(MaskTestMixin, DataOpsTestBase):
     # -- Null edge case tests ------------------------------------------------
 
     def test_null_policy_skip_all_null_column(self) -> None:
-        """NullPolicy.SKIP: score column is all null. Aggregation should produce all nulls."""
+        """Null-skip: score column is all null. Aggregation should produce all nulls."""
         fs = make_feature_set("score__sum_agg", ["region"])
         result = self.implementation_class().calculate_feature(self.test_data, fs)
