+++ mloda/testing/feature_groups/data_operations/row_preserving/ema/ema.py
@@ -250,15 +250,6 @@ class EmaTestBase(ReservedColumnsTestMixin, DataOpsTestBase):
 
     # -- Helpers ------------------------------------------------------------
 
-    def _assert_float_list_with_nulls(self, actual: list[Any], expected: list[Any]) -> None:
-        assert len(actual) == len(expected), f"row count {len(actual)} != expected {len(expected)}"
-        for i, (a, e) in enumerate(zip(actual, expected)):
-            if e is None:
-                assert a is None, f"row {i}: expected None, got {a!r}"
-            else:
-                assert a is not None, f"row {i}: expected {e!r}, got None"
-                assert float(a) == pytest.approx(e), f"row {i}: {a!r} != {e!r}"
-
     def _ema_feature_set(self, span: int) -> FeatureSet:
         return make_feature_set(f"value__ema_{span}", partition_by=["region"], order_by="ts")
