+++ mloda/testing/feature_groups/data_operations/row_preserving/window_aggregation/window_aggregation.py
@@ -208,7 +208,7 @@ class WindowAggregationTestBase(ReservedColumnsTestMixin, MaskTestMixin, DataOps
         assert result_col == EXPECTED_MAX_BY_REGION
 
     def test_null_policy_skip_avg_with_null_values(self) -> None:
-        """NullPolicy.SKIP: Group B has a null value_int at row 4. Avg should skip it."""
+        """Null-skip: Group B has a null value_int at row 4. Avg should skip it."""
         fs = make_feature_set("value_int__avg_window", ["region"])
         result = self.implementation_class().calculate_feature(self.test_data, fs)
 
@@ -219,7 +219,7 @@ class WindowAggregationTestBase(ReservedColumnsTestMixin, MaskTestMixin, DataOps
         assert result_col[7] == pytest.approx(GROUP_B_AVG_EXPECTED, rel=1e-6)
 
     def test_null_policy_null_is_group(self) -> None:
-        """NullPolicy.NULL_IS_GROUP: Row 11 has region=None. It should form its own group."""
+        """Null-as-group: Row 11 has region=None. It should form its own group."""
         fs = make_feature_set("value_int__sum_window", ["region"])
         result = self.implementation_class().calculate_feature(self.test_data, fs)
 
@@ -463,7 +463,7 @@ class WindowAggregationTestBase(ReservedColumnsTestMixin, MaskTestMixin, DataOps
     # -- Null edge case tests ------------------------------------------------
 
     def test_null_policy_skip_all_null_column(self) -> None:
-        """NullPolicy.SKIP: score column is all null. Aggregation should produce all nulls or zeros.
+        """Null-skip: score column is all null. Aggregation should produce all nulls or zeros.
 
         PyArrow/Polars/DuckDB/SQLite return null for sum of all-null group.
         Pandas returns 0 (groupby.transform("sum") treats all-null as 0).
